+++ launcher.py
@@ -1,5 +1,5 @@
 from lib.bot import bot
 
-VERSION = "2.8.0"
+VERSION = "2.8.1"
 
 bot.run(VERSION)